pyproject.toml
[project]
name = "mcp-mysql"
version = "0.1.0"
description = "MCP server for MySQL database connections and queries"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=0.9.0",
    "mysql-connector-python>=8.0.0",
    "starlette>=0.27.0",
    "uvicorn>=0.23.0",
    "sse-starlette>=1.6.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
mcp-mysql = "mcp_mysql.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
